+++ dart/constraint/ContactConstraint.cpp
@@ -63,6 +63,9 @@ double ContactConstraint::mErrorReductionParameter = DART_ERP;
 double ContactConstraint::mMaxErrorReductionVelocity = DART_MAX_ERV;
 double ContactConstraint::mConstraintForceMixing = DART_CFM;
 
+constexpr double DART_DEFAULT_FRICTION_COEFF = 1.0;
+constexpr double DART_DEFAULT_RESTITUTION_COEFF = 0.0;
+
 //==============================================================================
 ContactConstraint::ContactConstraint(
     collision::Contact& contact, double timeStep)
@@ -88,11 +91,19 @@ ContactConstraint::ContactConstraint(
   assert(
       contact.normal.squaredNorm() >= DART_CONTACT_CONSTRAINT_EPSILON_SQUARED);
 
+  const auto* shapeNodeA = const_cast<dynamics::ShapeFrame*>(
+                               contact.collisionObject1->getShapeFrame())
+                               ->asShapeNode();
+  const auto* shapeNodeB = const_cast<dynamics::ShapeFrame*>(
+                               contact.collisionObject2->getShapeFrame())
+                               ->asShapeNode();
+
   //----------------------------------------------
   // Bounce
   //----------------------------------------------
-  mRestitutionCoeff
-      = mBodyNodeA->getRestitutionCoeff() * mBodyNodeB->getRestitutionCoeff();
+  const double restitutionCoeffA = computeRestitutionCoefficient(shapeNodeA);
+  const double restitutionCoeffB = computeRestitutionCoefficient(shapeNodeB);
+  mRestitutionCoeff = restitutionCoeffA * restitutionCoeffB;
   if (mRestitutionCoeff > DART_RESTITUTION_COEFF_THRESHOLD)
     mIsBounceOn = true;
   else
@@ -104,8 +115,12 @@ ContactConstraint::ContactConstraint(
   // TODO(JS): Assume the frictional coefficient can be changed during
   //           simulation steps.
   // Update mFrictionalCoff
-  mFrictionCoeff = std::min(
-      mBodyNodeA->getFrictionCoeff(), mBodyNodeB->getFrictionCoeff());
+  const double frictionCoeffA = computeFrictionCoefficient(shapeNodeA);
+  const double frictionCoeffB = computeFrictionCoefficient(shapeNodeB);
+
+  // TODO(JS): Consider providing various ways of the combined friction or
+  // allowing to override this method by a custom method
+  mFrictionCoeff = std::min(frictionCoeffA, frictionCoeffB);
   if (mFrictionCoeff > DART_FRICTION_COEFF_THRESHOLD)
   {
     mIsFrictionOn = true;
@@ -659,6 +674,46 @@ void ContactConstraint::getRelVelocity(double* relVel)
 bool ContactConstraint::isActive() const
 {
   return mActive;
+}
+
+//==============================================================================
+double ContactConstraint::computeFrictionCoefficient(
+    const dynamics::ShapeNode* shapeNode)
+{
+  assert(shapeNode);
+
+  auto dynamicAspect = shapeNode->getDynamicsAspect();
+
+  if (dynamicAspect == nullptr)
+  {
+    dtwarn << "[ContactConstraint] Attempt to extract friction coefficient "
+           << "from a ShapeNode that doesn't have DynamicAspect. The default "
+           << "value (" << DART_DEFAULT_FRICTION_COEFF << ") will be used "
+           << "instead.\n";
+    return DART_DEFAULT_FRICTION_COEFF;
+  }
+
+  return dynamicAspect->getFrictionCoeff();
+}
+
+//==============================================================================
+double ContactConstraint::computeRestitutionCoefficient(
+    const dynamics::ShapeNode* shapeNode)
+{
+  assert(shapeNode);
+
+  auto dynamicAspect = shapeNode->getDynamicsAspect();
+
+  if (dynamicAspect == nullptr)
+  {
+    dtwarn << "[ContactConstraint] Attempt to extract restitution coefficient "
+           << "from a ShapeNode that doesn't have DynamicAspect. The default "
+           << "value (" << DART_DEFAULT_RESTITUTION_COEFF << ") will be used "
+           << "instead.\n";
+    return DART_DEFAULT_RESTITUTION_COEFF;
+  }
+
+  return dynamicAspect->getRestitutionCoeff();
 }
 
 //==============================================================================